Refactor merge commit
