+++ app/build.gradle
@@ -54,6 +54,7 @@ dependencies {
     implementation 'com.google.firebase:firebase-auth-ktx'
     implementation 'com.firebaseui:firebase-ui-auth:7.2.0'
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-play-services:1.3.5'
+
     /** fragment **/
     def fragment_version = "1.4.1"
     implementation "androidx.fragment:fragment-ktx:$fragment_version"
