--- app/src/main/java/com/example/hikeculator/data/firestore_constants/Constants.kt
@@ -1,3 +1,0 @@
-package com.example.hikeculator.data.firestore_constants
-
-const val USER_COLLECTION_NAME = "users"
